Return Err when cookie value is empty string
Destroying a session by logging out sets the cookie header to expire
when the browser session ends. It also sets the session ID to an empty
string.

The session-grant endopoint however sees the empty string as a valid
session key then tries to retrive that session which it never finds.

To get around this i'm returning a Err when parsing the cookie that has
an empty string value which initiates the github Auth flow again.