Set ExhibitView
Because we are using the same page for signed in authors and visitors
its handy to know the status of the person viewing the page.

If the person viewing is the author of the exhibits then we can load
views specific for them such as creating new exhibits etc.

A Visitor can only view the exhibits that are loaded and nothing else.