Add function to calculate the next exhibits view context
We are rendering custom views depending on whether the user is signed in
or not. If a viewer is signed in and is the Uthor of the exhibits, that
user will be able to do admin level operations like create new exhibits.

If they are not signed in or are not the author of the exhibits they
will only be able to view the exhibits as a visitor.

An alternative solution is to render a completely different page which
will decouple things but will mean lots of replication in the view.

I may regret this decision.